Moved data types out of Helpers, to Data.hs
