+++ code/drasil-code/drasil-code.cabal
@@ -16,6 +16,7 @@ library
     , Language.Drasil.Code.Imperative.Build.AST
     , Language.Drasil.Code.Imperative.Build.Import
     , Language.Drasil.Code.Imperative.Helpers
+    , Language.Drasil.Code.Imperative.Data
     , Language.Drasil.Code.Imperative.Import
     , Language.Drasil.Code.Imperative.LanguageRenderer
     , Language.Drasil.Code.Imperative.LanguageRenderer.JavaRenderer
@@ -61,6 +62,7 @@ executable codegenTest
     , Language.Drasil.Code.Imperative.Build.AST
     , Language.Drasil.Code.Imperative.Build.Import
     , Language.Drasil.Code.Imperative.Helpers
+    , Language.Drasil.Code.Imperative.Data
     , Language.Drasil.Code.Imperative.Import
     , Language.Drasil.Code.Imperative.LanguageRenderer
     , Language.Drasil.Code.Imperative.LanguageRenderer.JavaRenderer
